move set_mock_args from global_vars.py to data/__init__.py
