+++ megatron/tests/unit_tests/data/test_builder.py
@@ -98,7 +98,7 @@ def test_builder():
         def __getitem__(self, idx: int) -> Dict[str, numpy.ndarray]:
             return {"text": self.dataset[self.sample_index[idx]]}
 
-    from megatron.training.global_vars import set_mock_args
+    from tests.unit_tests.data import set_mock_args
     set_mock_args()
 
     with tempfile.TemporaryDirectory() as temp_dir:
